refactor: move types into type file
